# Conflicts:
#	app/src/main/java/org/openedx/app/AppRouter.kt
#	app/src/main/java/org/openedx/app/di/ScreenModule.kt
#	auth/src/main/java/org/openedx/auth/presentation/signin/SignInFragment.kt
#	auth/src/main/java/org/openedx/auth/presentation/signin/SignInViewModel.kt
#	auth/src/main/java/org/openedx/auth/presentation/signin/compose/SignInView.kt
#	build.gradle
#	core/src/main/java/org/openedx/core/config/Config.kt
#	core/src/main/res/values/strings.xml
#	default_config/prod/config.yaml
#	default_config/stage/config.yaml

+++ core/src/main/java/org/openedx/core/module/download/AbstractDownloader.kt
@@ -35,40 +35,45 @@ abstract class AbstractDownloader : KoinComponent {
     open suspend fun download(
         url: String,
         path: String
-    ): Boolean {
+    ): DownloadResult {
         isCanceled = false
         return try {
-            val response = downloadApi.downloadFile(url).body()
-            if (response != null) {
-                val file = File(path)
-                if (file.exists()) {
-                    file.delete()
+            val responseBody = downloadApi.downloadFile(url).body() ?: return DownloadResult.ERROR
+            initializeFile(path)
+            responseBody.byteStream().use { inputStream ->
+                FileOutputStream(File(path)).use { outputStream ->
+                    writeToFile(inputStream, outputStream)
                 }
-                file.createNewFile()
-                input = response.byteStream()
-                currentDownloadingFilePath = path
-                fos = FileOutputStream(file)
-                fos.use { output ->
-                    val buffer = ByteArray(4 * 1024)
-                    var read: Int
-                    while (input!!.read(buffer).also { read = it } != -1) {
-                        output?.write(buffer, 0, read)
-                    }
-                    output?.flush()
-                }
-                true
-            } else {
-                false
             }
+            DownloadResult.SUCCESS
         } catch (e: Exception) {
             e.printStackTrace()
-            false
+            if (isCanceled) DownloadResult.CANCELED else DownloadResult.ERROR
         } finally {
-            fos?.close()
-            input?.close()
+            closeResources()
         }
     }
 
+    private fun initializeFile(path: String) {
+        val file = File(path)
+        if (file.exists()) file.delete()
+        file.createNewFile()
+        currentDownloadingFilePath = path
+    }
+
+    private fun writeToFile(inputStream: InputStream, outputStream: FileOutputStream) {
+        val buffer = ByteArray(BUFFER_SIZE)
+        var bytesRead: Int
+        while (inputStream.read(buffer).also { bytesRead = it } != -1) {
+            outputStream.write(buffer, 0, bytesRead)
+        }
+        outputStream.flush()
+    }
+
+    private fun closeResources() {
+        fos?.close()
+        input?.close()
+    }
 
     suspend fun cancelDownloading() {
         isCanceled = true
@@ -88,4 +93,11 @@ abstract class AbstractDownloader : KoinComponent {
         }
     }
 
+    enum class DownloadResult {
+        SUCCESS, CANCELED, ERROR
+    }
+
+    companion object {
+        private const val BUFFER_SIZE = 4 * 1024
+    }
 }
